requirements.txt
PyYAML==6.0.2            
joblib==1.4.2           
# torch==2.6.0+cu118 #optional, uncomment if you want to use train models on GPU
# torchvision==0.21.0+cu118 #optional, uncomment if you want to use train models on GPU
torch
torchvision
numpy==2.1.1            
Pillow==11.0.0          
pyTelegramBotAPI==4.26.0 
pandas==2.2.3           
scikit-learn==1.6.1     
tqdm==4.67.1            
pycocotools==2.0.8       
matplotlib==3.10.1
ipywidgets==8.1.6
opencv-python==4.11.0.86